Add custom message to new error class
The evaluation name returned from google doesn't contain the name of the sample
query set used for the evalution. Add this information to the error message so that
when the error is raised in Sentry we can see which sample query set has not been evaluated
and decide if we want to re-run it.
https://stackoverflow.com/questions/11636874/raise-custom-exception-with-arguments