Configure automatic merging of some dependencies
Configure dependabot to allow automatic patching
of patch and minor dependencies, in line with
guidance from RFC-167 [^1]. It has been confirmed
that the repo meets the conditions of this RFC.

Rails and google-cloud-discovery_engine-v1beta have
been configured to only allow patch dependencies.
This follows usual guidance for Rails [^2], and
since this application has a very heavy dependency
on google-cloud-discovery-engine-v1beta, this
approach seems prudent.

[^1]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-167-auto-patch-dependencies.md
[^2]: https://docs.publishing.service.gov.uk/manual/keeping-software-current.html#rails